refactor: extract asyncWalk to lib/async-walk.js
Separate generic directory walking (lib/async-walk.js) from
config-array integration (index.js). Each file now has a single
responsibility. asyncWalk is re-exported from the main entry point.
