+++ index.js
@@ -1,5 +1,8 @@
-import { opendir, stat } from 'node:fs/promises';
-import path from 'node:path';
+import { stat } from 'node:fs/promises';
+
+import { asyncWalk as asyncWalkImpl } from './lib/async-walk.js';
+
+export { asyncWalk } from './lib/async-walk.js';
 
 /**
  * @typedef {object} ConfigLoader
@@ -7,78 +10,8 @@ import path from 'node:path';
  * @property {(filePath: string) => object | undefined | Promise<object | undefined>} getConfig Get config for a file. Returns undefined if file has no matching config.
  */
 
-/**
- * @typedef {object} WalkEntry
- * @property {string} path The full path of the entry.
- * @property {import('node:fs').Dirent} dirent The directory entry.
- */
-
-/**
- * @typedef {object} AsyncWalkOptions
- * @property {string} basePath The directory to walk.
- * @property {(entry: WalkEntry) => boolean | Promise<boolean>} deepFilter Filter for directory traversal.
- * @property {(entry: WalkEntry) => boolean | Promise<boolean>} entryFilter Filter for file inclusion.
- * @property {boolean} [followSymbolicLinks] Whether to follow symbolic links.
- * @property {AbortSignal} [signal] An AbortSignal to cancel the traversal.
- * @property {(error: NodeJS.ErrnoException) => boolean} [errorFilter] Optional function to filter errors. Return true to skip the error.
- */
-
-/**
- * Recursively walks a directory tree, applying async filter functions.
- *
- * @param {AsyncWalkOptions} options
- * @returns {Promise<string[]>} An array of matching file paths.
- */
-export async function asyncWalk (options) {
-  const { basePath, deepFilter, entryFilter, errorFilter, followSymbolicLinks, signal } = options;
-  /** @type {string[]} */
-  const results = [];
-
-  /**
-   * @param {string} dirPath
-   * @returns {Promise<void>}
-   */
-  async function walk (dirPath) {
-    signal?.throwIfAborted();
-
-    /** @type {import('node:fs').Dir | undefined} */
-    let dir;
-
-    try {
-      dir = await opendir(dirPath);
-    } catch (err) {
-      if (errorFilter?.(/** @type {NodeJS.ErrnoException} */ (err))) return;
-      throw err;
-    }
-
-    try {
-      for await (const dirent of dir) {
-        const fullPath = path.join(dirPath, dirent.name);
-
-        /** @type {WalkEntry} */
-        const entry = { path: fullPath, dirent };
-
-        // eslint-disable-next-line security/detect-non-literal-fs-filename -- fullPath is derived from walked directory
-        const resolvedStat = followSymbolicLinks && dirent.isSymbolicLink() ? await stat(fullPath).catch(() => {}) : undefined;
-        const isDir = dirent.isDirectory() || resolvedStat?.isDirectory();
-
-        if (isDir) {
-          if (await deepFilter(entry)) {
-            await walk(fullPath);
-          }
-        } else if (await entryFilter(entry)) {
-          results.push(fullPath);
-        }
-      }
-    } finally {
-      await dir.close().catch(() => {});
-    }
-  }
-
-  await walk(basePath);
-
-  return results.sort();
-}
+/** @typedef {import('./lib/async-walk.js').WalkEntry} WalkEntry */
+/** @typedef {import('./lib/async-walk.js').AsyncWalkOptions} AsyncWalkOptions */
 
 /**
  * Wraps a ConfigArray as a ConfigLoader.
@@ -103,8 +36,8 @@ export function configsToLoader (configs) {
  * @param {string} options.basePath The directory to search.
  * @param {import('@eslint/config-array').ConfigArray} [options.configs] The config array to use for determining what to ignore.
  * @param {ConfigLoader} [options.configLoader] A config loader with async-capable isDirectoryIgnored/getConfig methods. Alternative to configs.
- * @param {(entry: WalkEntry) => boolean} [options.deepFilter] Optional function that indicates whether the directory will be read deep or not.
- * @param {(entry: WalkEntry) => boolean} [options.entryFilter] Optional function that indicates whether the entry will be included to results or not.
+ * @param {(entry: import('./lib/async-walk.js').WalkEntry) => boolean} [options.deepFilter] Optional function that indicates whether the directory will be read deep or not.
+ * @param {(entry: import('./lib/async-walk.js').WalkEntry) => boolean} [options.entryFilter] Optional function that indicates whether the entry will be included to results or not.
  * @param {boolean} [options.followSymbolicLinks] Follow symbolic links when walking directories. Default: false.
  * @param {AbortSignal} [options.signal] An AbortSignal to cancel the traversal.
  * @param {(error: NodeJS.ErrnoException) => boolean} [options.errorFilter] Optional function to filter errors during traversal. Return true to skip the error and continue.
@@ -135,7 +68,7 @@ export async function configArrayFindFiles (options) {
 
   const loader = configLoader || configsToLoader(/** @type {import('@eslint/config-array').ConfigArray} */ (configs));
 
-  return asyncWalk({
+  return asyncWalkImpl({
     basePath,
     deepFilter: async (entry) => {
       if (deepFilter && !deepFilter(entry)) {
